# perf: Fanout for async updates in global