--- a/config.go
+++ b/config.go
@@ -64,6 +64,9 @@ type BehaviorConfig struct {
 	GlobalBatchLimit int
 	// ForceGlobal forces global mode on all rate limit checks.
 	ForceGlobal bool
+
+	// Number of concurrent requests that will be made to peers. Defaults to 100
+	PeerRequestsConcurrency int
 }
 
 // Config for a gubernator instance
@@ -125,6 +128,8 @@ func (c *Config) SetDefaults() error {
 	setter.SetDefault(&c.Behaviors.GlobalTimeout, time.Millisecond*500)
 	setter.SetDefault(&c.Behaviors.GlobalBatchLimit, maxBatchSize)
 	setter.SetDefault(&c.Behaviors.GlobalSyncWait, time.Millisecond*100)
+
+	setter.SetDefault(&c.Behaviors.PeerRequestsConcurrency, 100)
 
 	setter.SetDefault(&c.LocalPicker, NewReplicatedConsistentHash(nil, defaultReplicas))
 	setter.SetDefault(&c.RegionPicker, NewRegionPicker(nil))